gui refactor wip
